feat: move `CreateSuggestionWithRecordId` class
